Wrap R to enable 'backcasting' validation
Call out to R so that helper models can be programmatically trained on
different subsets of data. This allows us to validate our model with
backcasting - projecting historic periods - so that we can compare our
model projections to the historic actuals. We also train a simple
linear regression model on the same subsets of data for an additional comparison.